Fix issue with crash on empty stage

See merge request Isaac/omni_isaac_sim!1143

+++ source/extensions/omni.isaac.ros_ui/python/scripts/roscore.py
@@ -2,17 +2,18 @@ import subprocess
 import sys
 import signal
 import os
+import psutil
 
 
 def kill_child_processes(parent_pid, sig=signal.SIGTERM):
     try:
-        os.kill(parent_pid, 0)
-        print(parent_pid)
-    except OSError:
-        print("parent process not existing")
+        parent = psutil.Process(parent_pid)
+    except psutil.NoSuchProcess:
         return
-    print("try to kill child: " + str(parent_pid))
-    os.kill(parent_pid, sig)
+    children = parent.children(recursive=True)
+    for process in children:
+        print("killing child: ", process)
+        process.send_signal(sig)
 
 
 class Roscore(object):
@@ -27,9 +28,26 @@ class Roscore(object):
         if Roscore.__initialized:
             raise Exception("You can't create more than 1 instance of Roscore.")
         Roscore.__initialized = True
+        self.roscore_process = None
+        self.roscore_pid = None
+
+    def check_running(self):
+        import rosgraph
+
+        try:
+            rosgraph.Master("/rostopic").getPid()
+        except:
+            print("ROS master is not running")
+            return False
+        else:
+            print("ROS master is already running")
+            return True
 
     def startup(self, python_path, ros_path, prefix):
         ros_env = {}
+        if self.check_running() == True:
+            print("Not starting internal roscore")
+            return
         # Get ros env parameters from bash setup script
         try:
             self.bash_process = subprocess.check_output(
@@ -46,16 +64,19 @@ class Roscore(object):
 
         try:
             # append path to specified python bin folder
-            ros_env["PATH"] = f"{python_path}:" + ros_env["PATH"]
-            self.roscore_process = subprocess.Popen(["roscore"], cwd=f"{ros_path}", env=ros_env)
+            ros_env["PATH"] = f"{python_path}:{ros_path}/bin:/bin"
+            # + ros_env["PATH"]
+            # print(ros_env)
+            self.roscore_process = subprocess.Popen(["roscore"], shell=True, cwd=f"{ros_path}", env=ros_env)
             self.roscore_pid = self.roscore_process.pid  # pid of the roscore process (which has child processes)
         except OSError as e:
             sys.stderr.write("roscore could not be run")
             raise e
 
     def shutdown(self):
-        print("try to kill child pids of roscore pid: " + str(self.roscore_pid))
-        kill_child_processes(self.roscore_pid)
-        self.roscore_process.terminate()
-        self.roscore_process.wait()  # important to prevent from zombie process
+        if self.roscore_pid is not None and self.roscore_process is not None:
+            print("try to kill child pids of roscore pid: " + str(self.roscore_pid))
+            kill_child_processes(self.roscore_pid)
+            self.roscore_process.terminate()
+            self.roscore_process.wait()  # important to prevent from zombie process
         Roscore.__initialized = False
